git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/branches/solr_3159_jetty8@1297628 13f79535-47bb-0310-9956-ffa450edef68

+++ modules/suggest/src/test/org/apache/lucene/search/suggest/TermFreq.java
@@ -21,13 +21,13 @@ import org.apache.lucene.util.BytesRef;
 
 public final class TermFreq {
   public final BytesRef term;
-  public final float v;
+  public final long v;
 
-  public TermFreq(String term, float v) {
+  public TermFreq(String term, long v) {
    this(new BytesRef(term), v);
   }
   
-  public TermFreq(BytesRef term, float v) {
+  public TermFreq(BytesRef term, long v) {
     this.term = term;
     this.v = v;
   }
